requirements.txt
aiogram==3.4.1
sqlalchemy==2.0.23
aiosqlite==0.19.0
pydantic==2.5.2
python-dotenv==1.0.0
alembic==1.13.1
redis==5.0.1
aiogram[redis]==3.4.1
